pyproject.toml
[tool.poetry]
name = "transskribilo-boto"
version = "0.1.0"
description = "Roboto por aŭtomata transskribado el x-sistemo al unikodan tekston."
authors = ["Marco Cognetta <cognetta.marco@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "3.7.4.post0"
async-timeout = "3.0.1"
attrs = "21.2.0"
chardet = "4.0.0"
"discord.py" = "1.7.3"
idna = "3.2"
multidict = "5.1.0"
typing-extensions = "3.10.0.0"
yarl = "1.6.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
